requirements.txt
streamlit
google-generativeai
openai
PyPDF2
